diff-engine/requirements.txt
deepdiff==8.2.0
rapidfuzz==3.10.1
redis==5.2.1
psycopg2-binary==2.9.10
xlsxwriter==3.2.2
jinja2==3.1.5
